+++ include/qpp/classes/qcircuit.hpp
@@ -57,18 +57,17 @@
 #include "qpp/classes/idisplay.hpp"
 #include "qpp/classes/ijson.hpp"
 
-#include "qpp/internal/classes/qcircuit_conditional_step.hpp"
 #include "qpp/internal/classes/qcircuit_gate_step.hpp"
 #include "qpp/internal/classes/qcircuit_measurement_step.hpp"
 #include "qpp/internal/classes/qcircuit_nop_step.hpp"
 #include "qpp/internal/classes/qcircuit_resources.hpp"
+#include "qpp/internal/classes/qcircuit_runtime_step.hpp"
 #include "qpp/internal/util.hpp"
 
 namespace qpp {
 namespace internal {
 ///< Conditional stack type
-using conditional_stack_t =
-    std::deque<internal::QCircuitConditionalStep::Context>;
+using conditional_stack_t = std::deque<internal::QCircuitRuntimeStep::Context>;
 
 ///< Forward declarations
 class QCircuitIterator;
@@ -108,7 +107,7 @@ class QCircuit : public IDisplay, public IJSON {
         measurement_count_{}; ///< measurement counts
 
     internal::conditional_stack_t
-        conditional_stack_{}; ///< used for parsing conditional statements
+        conditional_stack_{}; ///< to parse conditional (runtime) statements
     std::optional<idx>
         outer_while_pos_{}; ///< location of the outermost while statement
 
@@ -137,9 +136,10 @@ class QCircuit : public IDisplay, public IJSON {
     }
 
     ///< circuit step variant
-    using VarStep = std::variant<
-        internal::QCircuitConditionalStep, internal::QCircuitGateStep,
-        internal::QCircuitMeasurementStep, internal::QCircuitNOPStep>;
+    using VarStep =
+        std::variant<internal::QCircuitRuntimeStep, internal::QCircuitGateStep,
+                     internal::QCircuitMeasurementStep,
+                     internal::QCircuitNOPStep>;
     std::vector<VarStep> circuit_{}; ///<< quantum circuit representation
 
   protected:
@@ -746,7 +746,7 @@ class QCircuit : public IDisplay, public IJSON {
     }
 
     /*
-     * \brief Validates the conditional statements
+     * \brief Validates the conditional if/else/while (runtime) statements
      *
      * \return True if valid conditionals, false otherwise
      */
@@ -814,10 +814,10 @@ class QCircuit : public IDisplay, public IJSON {
                 }
             };
         auto nop_step_visitor_id = [&](internal::QCircuitNOPStep&) {};
-        auto conditional_step_visitor =
-            [&](const internal::QCircuitConditionalStep&) {};
+        auto runtime_step_visitor = [&](const internal::QCircuitRuntimeStep&) {
+        };
 
-        visit_circuit_(circuit_, conditional_step_visitor, gate_step_visitor,
+        visit_circuit_(circuit_, runtime_step_visitor, gate_step_visitor,
                        measurement_step_visitor, nop_step_visitor_id);
 
         // update the destructively measured qudits
@@ -890,16 +890,16 @@ class QCircuit : public IDisplay, public IJSON {
                 }
             };
         // update condition statement context indexes
-        auto conditional_step_visitor =
-            [&](internal::QCircuitConditionalStep& conditional_step) {
+        auto runtime_step_visitor =
+            [&](internal::QCircuitRuntimeStep& runtime_step) {
                 // add dit context (offset, length)
-                if (conditional_step.ctx_.start_expr.has_value()) {
-                    conditional_step.ctx_.dit_ctx.emplace_back(pos, n);
+                if (runtime_step.ctx_.start_expr.has_value()) {
+                    runtime_step.ctx_.dit_ctx.emplace_back(pos, n);
                 }
             };
         auto nop_step_visitor_id = [&](internal::QCircuitNOPStep&) {};
 
-        visit_circuit_(circuit_, conditional_step_visitor, gate_step_visitor,
+        visit_circuit_(circuit_, runtime_step_visitor, gate_step_visitor,
                        measurement_step_visitor, nop_step_visitor_id);
 
         // update (enlarge) the clean dits vector
@@ -969,11 +969,11 @@ class QCircuit : public IDisplay, public IJSON {
             outer_while_pos_ = get_step_count();
         }
 
-        internal::QCircuitConditionalStep::Context ctx{};
+        internal::QCircuitRuntimeStep::Context ctx{};
         ctx.start_expr = {get_step_count(), pred};
         conditional_stack_.push_back(ctx);
-        circuit_.emplace_back(internal::QCircuitConditionalStep{
-            internal::QCircuitConditionalStep::Type::WHILE, ctx});
+        circuit_.emplace_back(internal::QCircuitRuntimeStep{
+            internal::QCircuitRuntimeStep::Type::WHILE, ctx});
 
         measured_d_.push_back(measured_d_.back());
 
@@ -1022,11 +1022,11 @@ class QCircuit : public IDisplay, public IJSON {
      * \return Reference to the current instance
      */
     QCircuit& cond_if(cond_pred_t pred) {
-        internal::QCircuitConditionalStep::Context ctx{};
+        internal::QCircuitRuntimeStep::Context ctx{};
         ctx.start_expr = {get_step_count(), pred};
         conditional_stack_.push_back(ctx);
-        circuit_.emplace_back(internal::QCircuitConditionalStep{
-            internal::QCircuitConditionalStep::Type::IF, ctx});
+        circuit_.emplace_back(internal::QCircuitRuntimeStep{
+            internal::QCircuitRuntimeStep::Type::IF, ctx});
 
         measured_d_.push_back(measured_d_.back());
 
@@ -1049,20 +1049,18 @@ class QCircuit : public IDisplay, public IJSON {
                 "qpp::QCircuit::cond_else()",
                 context + ": ELSE without a matching IF");
         }
-        if (std::get<internal::QCircuitConditionalStep>(
+        if (std::get<internal::QCircuitRuntimeStep>(
                 circuit_[conditional_stack_.back().start_expr->first])
-                .condition_type_ ==
-            internal::QCircuitConditionalStep::Type::WHILE) {
+                .runtime_type_ == internal::QCircuitRuntimeStep::Type::WHILE) {
             throw exception::InvalidConditional("qpp::QCircuit::cond_else()",
                                                 context + ": WHILE ELSE");
         }
         // END EXCEPTION CHECKS
 
-        internal::QCircuitConditionalStep::Context& ctx =
-            conditional_stack_.back();
+        internal::QCircuitRuntimeStep::Context& ctx = conditional_stack_.back();
         ctx.else_expr = get_step_count();
-        circuit_.emplace_back(internal::QCircuitConditionalStep{
-            internal::QCircuitConditionalStep::Type::ELSE, ctx});
+        circuit_.emplace_back(internal::QCircuitRuntimeStep{
+            internal::QCircuitRuntimeStep::Type::ELSE, ctx});
 
         measured_d_if_stack_.push_back(measured_d_.back());
         measured_d_.pop_back();
@@ -1088,32 +1086,29 @@ class QCircuit : public IDisplay, public IJSON {
         }
         // END EXCEPTION CHECKS
 
-        internal::QCircuitConditionalStep::Context& ctx =
-            conditional_stack_.back();
+        internal::QCircuitRuntimeStep::Context& ctx = conditional_stack_.back();
         ctx.end_expr = get_step_count();
-        if (std::get<internal::QCircuitConditionalStep>(
+        if (std::get<internal::QCircuitRuntimeStep>(
                 circuit_[ctx.start_expr->first])
-                .condition_type_ ==
-            internal::QCircuitConditionalStep::Type::WHILE) {
-            circuit_.emplace_back(internal::QCircuitConditionalStep{
-                internal::QCircuitConditionalStep::Type::ENDWHILE, ctx});
+                .runtime_type_ == internal::QCircuitRuntimeStep::Type::WHILE) {
+            circuit_.emplace_back(internal::QCircuitRuntimeStep{
+                internal::QCircuitRuntimeStep::Type::ENDWHILE, ctx});
             if (outer_while_pos_.value() == ctx.start_expr.value().first) {
                 outer_while_pos_ = std::nullopt;
             }
         } else {
-            circuit_.emplace_back(internal::QCircuitConditionalStep{
-                internal::QCircuitConditionalStep::Type::ENDIF, ctx});
+            circuit_.emplace_back(internal::QCircuitRuntimeStep{
+                internal::QCircuitRuntimeStep::Type::ENDIF, ctx});
         }
 
         // update the IF/ELSE contexts
         auto start_expr = ctx.start_expr;
         auto else_expr = ctx.else_expr;
         if (else_expr.has_value()) {
-            std::get<internal::QCircuitConditionalStep>(
-                circuit_[else_expr.value()])
+            std::get<internal::QCircuitRuntimeStep>(circuit_[else_expr.value()])
                 .ctx_ = ctx;
         }
-        std::get<internal::QCircuitConditionalStep>(
+        std::get<internal::QCircuitRuntimeStep>(
             circuit_[start_expr.value().first])
             .ctx_ = ctx;
         conditional_stack_.pop_back();
@@ -1150,10 +1145,10 @@ class QCircuit : public IDisplay, public IJSON {
      * \return Reference to the current instance
      */
     QCircuit& set_dits_runtime(mutable_dits_functor_t functor) {
-        internal::QCircuitConditionalStep::Context ctx{};
+        internal::QCircuitRuntimeStep::Context ctx{};
         ctx.set_dits_func = functor;
-        circuit_.emplace_back(internal::QCircuitConditionalStep{
-            internal::QCircuitConditionalStep::Type::SET_DITS_RUNTIME, ctx});
+        circuit_.emplace_back(internal::QCircuitRuntimeStep{
+            internal::QCircuitRuntimeStep::Type::SET_DITS_RUNTIME, ctx});
 
         return *this;
     }
@@ -4071,7 +4066,8 @@ class QCircuit : public IDisplay, public IJSON {
         add_dit(other.nc_, pos_dit.value());
 
         // STEP 1: update [c]ctrl and target indexes of other, and modify
-        // the locations in conditional steps + add dit context for other
+        // the locations in conditional (runtime) steps + add dit context for
+        // other
         auto gate_step_visitor = [&](internal::QCircuitGateStep& gate_step) {
             // update the cctrl indexes
             if (is_cCTRL(gate_step)) {
@@ -4103,20 +4099,20 @@ class QCircuit : public IDisplay, public IJSON {
                 }
             };
         auto nop_step_visitor_id = [&](internal::QCircuitNOPStep&) {};
-        auto conditional_step_visitor =
-            [&](internal::QCircuitConditionalStep& conditional_step) {
-                // increment location of conditional statements accordingly
-                conditional_step.ctx_.inc_cond_locs(get_step_count());
+        auto runtime_step_visitor =
+            [&](internal::QCircuitRuntimeStep& runtime_step) {
+                // increment location of conditional (runtime) statements
+                // accordingly
+                runtime_step.ctx_.inc_cond_locs(get_step_count());
                 // add dit context (offset, length) as if we're inserting
                 // classical dits into other before composing
-                auto& ctx_dit_vec = conditional_step.ctx_.dit_ctx;
+                auto& ctx_dit_vec = runtime_step.ctx_.dit_ctx;
                 ctx_dit_vec.insert(ctx_dit_vec.begin(),
                                    {0, pos_dit.value_or(this->get_nc())});
             };
 
-        visit_circuit_(other.circuit_, conditional_step_visitor,
-                       gate_step_visitor, measurement_step_visitor,
-                       nop_step_visitor_id);
+        visit_circuit_(other.circuit_, runtime_step_visitor, gate_step_visitor,
+                       measurement_step_visitor, nop_step_visitor_id);
 
         // STEP 2
         // replace the corresponding elements of measured_, measured_nd_,
@@ -4279,30 +4275,30 @@ class QCircuit : public IDisplay, public IJSON {
         // STEP 0: insert classical dits from the to-be-coupled circuit
         add_dit(other.nc_, pos_dit.value());
 
-        // STEP 1: modify the locations in conditional steps of current, and
-        // add dit context for other
+        // STEP 1: modify the locations in conditional (runtime) steps of
+        // current, and add dit context for other
         auto gate_step_visitor_id = [&](internal::QCircuitGateStep&) {};
         auto measurement_step_visitor_id =
             [&](internal::QCircuitMeasurementStep&) {};
         auto nop_step_visitor_id = [&](internal::QCircuitNOPStep&) {};
-        auto conditional_step_visitor_current =
-            [&](internal::QCircuitConditionalStep& conditional_step_current) {
-                // increment location of conditional statements accordingly
-                conditional_step_current.ctx_.inc_cond_locs(
-                    other.get_step_count());
+        auto runtime_step_visitor_current =
+            [&](internal::QCircuitRuntimeStep& runtime_step_current) {
+                // increment location of conditional (runtime) statements
+                // accordingly
+                runtime_step_current.ctx_.inc_cond_locs(other.get_step_count());
             };
-        auto conditional_step_visitor_other =
-            [&](internal::QCircuitConditionalStep& conditional_step_other) {
-                auto& ctx_dit_vec = conditional_step_other.ctx_.dit_ctx;
+        auto runtime_step_visitor_other =
+            [&](internal::QCircuitRuntimeStep& runtime_step_other) {
+                auto& ctx_dit_vec = runtime_step_other.ctx_.dit_ctx;
                 // add dit context (offset, length) as if we're inserting
                 // classical dits into other before composing
                 ctx_dit_vec.insert(ctx_dit_vec.begin(),
                                    {0, pos_dit.value_or(this->get_nc())});
             };
-        visit_circuit_(circuit_, conditional_step_visitor_current,
+        visit_circuit_(circuit_, runtime_step_visitor_current,
                        gate_step_visitor_id, measurement_step_visitor_id,
                        nop_step_visitor_id);
-        visit_circuit_(other.circuit_, conditional_step_visitor_other,
+        visit_circuit_(other.circuit_, runtime_step_visitor_other,
                        gate_step_visitor_id, measurement_step_visitor_id,
                        nop_step_visitor_id);
 
@@ -4336,10 +4332,10 @@ class QCircuit : public IDisplay, public IJSON {
                 }
             };
         auto other_nop_step_visitor_id = [&](internal::QCircuitNOPStep&) {};
-        auto other_conditional_step_visitor_id =
-            [&](const internal::QCircuitConditionalStep&) {};
+        auto other_runtime_step_visitor_id =
+            [&](const internal::QCircuitRuntimeStep&) {};
 
-        visit_circuit_(other.circuit_, other_conditional_step_visitor_id,
+        visit_circuit_(other.circuit_, other_runtime_step_visitor_id,
                        other_gate_step_visitor, other_measurement_step_visitor,
                        other_nop_step_visitor_id);
 
@@ -4505,7 +4501,7 @@ class QCircuit : public IDisplay, public IJSON {
             }
         };
         // update measurement indexes of other, modify the locations in
-        // conditional steps + add dit context for other
+        // conditional (runtime) steps + add dit context for other
         auto measurement_step_visitor =
             [&](internal::QCircuitMeasurementStep& measurement_step) {
                 measurement_step.c_reg_ += pos_dit.value();
@@ -4514,20 +4510,20 @@ class QCircuit : public IDisplay, public IJSON {
                 }
             };
         auto nop_step_visitor_id = [&](internal::QCircuitNOPStep&) {};
-        auto conditional_step_visitor =
-            [&](internal::QCircuitConditionalStep& conditional_step) {
-                // increment location of conditional statements accordingly
-                conditional_step.ctx_.inc_cond_locs(get_step_count());
+        auto runtime_step_visitor =
+            [&](internal::QCircuitRuntimeStep& runtime_step) {
+                // increment location of conditional (runtime) statements
+                // accordingly
+                runtime_step.ctx_.inc_cond_locs(get_step_count());
                 // add dit context (offset, length) as if we're inserting
                 // classical dits into other before composing
-                auto& ctx_dit_vec = conditional_step.ctx_.dit_ctx;
+                auto& ctx_dit_vec = runtime_step.ctx_.dit_ctx;
                 ctx_dit_vec.insert(ctx_dit_vec.begin(),
                                    {0, pos_dit.value_or(this->get_nc())});
             };
 
-        visit_circuit_(other.circuit_, conditional_step_visitor,
-                       gate_step_visitor, measurement_step_visitor,
-                       nop_step_visitor_id);
+        visit_circuit_(other.circuit_, runtime_step_visitor, gate_step_visitor,
+                       measurement_step_visitor, nop_step_visitor_id);
 
         // replace the corresponding elements of measured_, measured_nd_,
         // clean_qudits_, clean_dits_, and measurement_dits_ with the ones
@@ -4824,10 +4820,10 @@ class QCircuit : public IDisplay, public IJSON {
             [&](const internal::QCircuitMeasurementStep&) {};
         // NOPs are left unchanged
         auto nop_step_visitor_id = [&](const internal::QCircuitNOPStep&) {};
-        auto conditional_step_visitor_id =
-            [&](const internal::QCircuitConditionalStep&) {};
+        auto runtime_step_visitor_id =
+            [&](const internal::QCircuitRuntimeStep&) {};
         for (idx i = end_ctrl_circuit; i < end_composed_circuit; ++i) {
-            visit_circuit_step_(circuit_[i], conditional_step_visitor_id,
+            visit_circuit_step_(circuit_[i], runtime_step_visitor_id,
                                 gate_step_visitor, measurement_step_visitor_id,
                                 nop_step_visitor_id);
         }
@@ -4864,15 +4860,15 @@ class QCircuit : public IDisplay, public IJSON {
 
     /**
      * \brief Returns true if the quantum circuit description contains
-     * conditionals, false otherwise
+     * runtime steps, false otherwise
      *
-     * \return True if the quantum circuit description contains conditionals,
+     * \return True if the quantum circuit description contains runtime steps,
      * false otherwise
      */
-    bool has_conditionals() const noexcept {
+    bool has_runtime_steps() const noexcept {
         return std::find_if(circuit_.cbegin(), circuit_.cend(), [](auto&& arg) {
-                   return std::holds_alternative<
-                       internal::QCircuitConditionalStep>(arg);
+                   return std::holds_alternative<internal::QCircuitRuntimeStep>(
+                       arg);
                }) != circuit_.end();
     }
 
@@ -4930,7 +4926,7 @@ class QCircuit : public IDisplay, public IJSON {
             throw exception::QuditAlreadyMeasured(
                 "qpp::QCircuit::adjoint()", "Current qpp::QCircuit instance");
         }
-        if (this->has_conditionals()) {
+        if (this->has_runtime_steps()) {
             throw exception::CircuitContainsConditionals(
                 "qpp::QCircuit::adjoint()", "Current qpp::QCircuit instance");
         }
@@ -4961,10 +4957,10 @@ class QCircuit : public IDisplay, public IJSON {
         auto measurement_step_visitor_id =
             [&](const internal::QCircuitMeasurementStep&) {};
         auto nop_step_visitor_id = [&](const internal::QCircuitNOPStep&) {};
-        auto conditional_step_visitor_id =
-            [&](const internal::QCircuitConditionalStep&) {};
+        auto runtime_step_visitor_id =
+            [&](const internal::QCircuitRuntimeStep&) {};
 
-        visit_circuit_(circuit_, conditional_step_visitor_id, gate_step_visitor,
+        visit_circuit_(circuit_, runtime_step_visitor_id, gate_step_visitor,
                        measurement_step_visitor_id, nop_step_visitor_id);
 
         return *this;
@@ -5134,10 +5130,10 @@ class QCircuit : public IDisplay, public IJSON {
                 }
             };
         auto nop_step_visitor_id = [&](const internal::QCircuitNOPStep&) {};
-        auto conditional_step_visitor_id =
-            [&](const internal::QCircuitConditionalStep&) {};
+        auto runtime_step_visitor_id =
+            [&](const internal::QCircuitRuntimeStep&) {};
 
-        visit_circuit_(circuit_, conditional_step_visitor_id, gate_step_visitor,
+        visit_circuit_(circuit_, runtime_step_visitor_id, gate_step_visitor,
                        measurement_step_visitor, nop_step_visitor_id);
 
         clean_qudits_.erase(std::next(clean_qudits_.begin(),
@@ -5181,10 +5177,10 @@ class QCircuit : public IDisplay, public IJSON {
                 }
             };
         auto nop_step_visitor_id = [&](const internal::QCircuitNOPStep&) {};
-        auto conditional_step_visitor_id =
-            [&](const internal::QCircuitConditionalStep&) {};
+        auto runtime_step_visitor_id =
+            [&](const internal::QCircuitRuntimeStep&) {};
 
-        visit_circuit_(circuit_, conditional_step_visitor_id, gate_step_visitor,
+        visit_circuit_(circuit_, runtime_step_visitor_id, gate_step_visitor,
                        measurement_step_visitor, nop_step_visitor_id);
 
         clean_dits_.erase(std::next(clean_dits_.begin(),
@@ -5525,14 +5521,14 @@ class QCircuitIterator {
                 [&](const internal::QCircuitNOPStep& nop_step) {
                     os << nop_step;
                 };
-            auto conditional_step_visitor =
-                [&](const internal::QCircuitConditionalStep& conditional_step) {
-                    os << conditional_step;
+            auto runtime_step_visitor =
+                [&](const internal::QCircuitRuntimeStep& runtime_step) {
+                    os << runtime_step;
                 };
 
             qc_ptr_->visit_circuit_step_(
-                qc_ptr_->circuit_[ip_], conditional_step_visitor,
-                gate_step_visitor, measurement_step_visitor, nop_step_visitor);
+                qc_ptr_->circuit_[ip_], runtime_step_visitor, gate_step_visitor,
+                measurement_step_visitor, nop_step_visitor);
 
             return os;
         }
@@ -5890,10 +5886,10 @@ inline std::string QCircuit::to_JSON(bool enclosed_in_curly_brackets) const {
         auto nop_step_visitor = [&](const internal::QCircuitNOPStep&) {
             result += std::string{"\"NOP\""} + "}";
         };
-        auto conditional_step_visitor =
-            [&](const internal::QCircuitConditionalStep& conditional_step) {
-                auto type = conditional_step.condition_type_;
-                auto ctx = conditional_step.ctx_;
+        auto runtime_step_visitor =
+            [&](const internal::QCircuitRuntimeStep& runtime_step) {
+                auto type = runtime_step.runtime_type_;
+                auto ctx = runtime_step.ctx_;
 
                 std::ostringstream ss;
                 std::string type_str, ctx_str;
@@ -5908,7 +5904,7 @@ inline std::string QCircuit::to_JSON(bool enclosed_in_curly_brackets) const {
                           "\"" + "}";
             };
 
-        visit_circuit_step_(elem.get_step(), conditional_step_visitor,
+        visit_circuit_step_(elem.get_step(), runtime_step_visitor,
                             gate_step_visitor, measurement_step_visitor,
                             nop_step_visitor);
     } // end for
@@ -6358,7 +6354,7 @@ inline QCircuit adjoint(QCircuit qc,
     if (qc.has_measurements()) {
         throw exception::QuditAlreadyMeasured("qpp::adjoint()", "qc");
     }
-    if (qc.has_conditionals()) {
+    if (qc.has_runtime_steps()) {
         throw exception::CircuitContainsConditionals("qpp::adjoint()", "qc");
     }
     // END EXCEPTION CHECKS
@@ -7270,16 +7266,16 @@ inline bool is_cCTRL(const QCircuit::iterator::value_type& elem) {
 inline bool is_cCTRL(QCircuit::iterator it) { return is_cCTRL(*it); }
 
 /**
- * \brief True if the quantum circuit step is a conditionalstep, false
+ * \brief True if the quantum circuit step is a runtime step, false
  * otherwise
  *
  * \param elem Quantum circuit step
- * \return True if the quantum circuit step is a conditional step, false
+ * \return True if the quantum circuit step is a runtime step, false
  * otherwise
  */
-inline bool is_conditional(const QCircuit::iterator::value_type& elem) {
+inline bool is_runtime_step(const QCircuit::iterator::value_type& elem) {
     auto step = elem.get_step();
-    if (std::holds_alternative<internal::QCircuitConditionalStep>(step)) {
+    if (std::holds_alternative<internal::QCircuitRuntimeStep>(step)) {
         return true;
     }
 
@@ -7287,15 +7283,15 @@ inline bool is_conditional(const QCircuit::iterator::value_type& elem) {
 }
 
 /**
- * \brief True if the quantum circuit iterator points to a conditional step,
- * false otherwise
+ * \brief True if the quantum circuit iterator points to a runtime step, false
+ * otherwise
  *
  * \param it Quantum circuit iterator
- * \return True if the quantum circuit iterator points to a conditional
- * step, false otherwise
+ * \return True if the quantum circuit iterator points to a runtime step, false
+ * otherwise
  */
-inline bool is_conditional(QCircuit::iterator it) {
-    return is_conditional(*it);
+inline bool is_runtime_step(QCircuit::iterator it) {
+    return is_runtime_step(*it);
 }
 /**
  * \brief Extracts ctrl, target, ps_vals, and c_reg vectors (in this order)
@@ -7483,7 +7479,7 @@ circuit_as_iterators(const QCircuit& qc) {
  * re-ordering is stable
  *
  * \note This function has no effect if the circuit description contains
- * conditional statements
+ * runtime statements
  *
  * \param start Quantum circuit iterator pointing to the first element
  * \param finish Quantum circuit iterator pointing to the last element (not
@@ -7497,11 +7493,11 @@ canonical_form(QCircuit::iterator start, QCircuit::iterator finish) {
     // NOTE: optimize conditionals if possible (probably not
     // straightforward)
     //
-    // if the circuit contains conditional statements, return the circuit as
+    // if the circuit contains runtime statements, return the circuit as
     // is, do not reorder it
-    auto first_conditional_it = std::find_if(
-        start, finish, [](auto&& elem) { return is_conditional(elem); });
-    if (first_conditional_it != finish) {
+    auto first_runtime_it = std::find_if(
+        start, finish, [](auto&& elem) { return is_runtime_step(elem); });
+    if (first_runtime_it != finish) {
         return circuit_as_iterators(start, finish);
     }
 
@@ -7562,9 +7558,9 @@ canonical_form(QCircuit::iterator start, QCircuit::iterator finish) {
  * Measurements]
  *
  * \note This function has no effect if the circuit description contains
- * conditional statements
+ * runtime statements
  *
- * \note This function does not push across conditional statements
+ * \note This function does not push across runtime statements
  *
  * \param qc Quantum circuit description
  * \return Quantum circuit canonical form represented as a vector of quantum
@@ -7575,8 +7571,8 @@ inline std::vector<QCircuit::iterator> canonical_form(const QCircuit& qc) {
 }
 
 /**
- * \brief Increments by \a i all conditional locations in the conditional
- * stack \a cs
+ * \brief Increments by \a i all conditional if/else/while locations in the
+ * conditional stack \a cs
  *
  * \param cs Conditional stack
  * \param i Non-negative integer
@@ -7589,19 +7585,20 @@ inline void inc_conditional_stack(conditional_stack_t& cs, idx i) {
 
 /**
  * \brief True if the qpp::QCircuit::iterator range [\a start, \a finish)
- * contains conditional statements, false otherwise
+ * contains runtime statements, false otherwise
  *
  * \param start Quantum circuit iterator pointing to the first element
  * \param finish Quantum circuit iterator pointing to the last element (not
  included)
  * \return True if the qpp::QCircuit::iterator range [\a start, \a finish)
- * contains conditional statements, false otherwise
+ * contains runtime statements, false otherwise
 
  */
-inline bool has_conditionals(QCircuit::iterator start,
-                             QCircuit::iterator finish) {
-    return std::find_if(start, finish,
-                        [](auto&& it) { return is_conditional(it); }) != finish;
+inline bool has_runtime_steps(QCircuit::iterator start,
+                              QCircuit::iterator finish) {
+    return std::find_if(start, finish, [](auto&& it) {
+               return is_runtime_step(it);
+           }) != finish;
 }
 
 } /* namespace internal */