+++ unit_tests/tests/classes/qcircuit.cpp
@@ -253,10 +253,10 @@ TEST(qpp_QCircuit_get_resources, AllTests) {}
 TEST(qpp_QCircuit_get_step_count, AllTests) {}
 
 /// BEGIN bool QCircuit::has_measurements() const noexcept
-TEST(qpp_QCircuit_has_conditionals, AllTests) {}
-
-/// BEGIN bool QCircuit::has_measurements() const noexcept
 TEST(qpp_QCircuit_has_measurements, AllTests) {}
+
+/// BEGIN bool QCircuit::has_runtime_steps() const noexcept
+TEST(qpp_QCircuit_has_runtime_steps, AllTests) {}
 
 /// BEGIN inline static bool QCircuit::is_cCTRL(const GateStep& gate_step)
 TEST(qpp_QCircuit_is_cCTRL, AllTests) {}
@@ -448,7 +448,7 @@ TEST(qpp_QCircuit_TFQ, SpecificQudits) {}
 TEST(qpp_QCircuit_to_JSON, AllTests) {}
 
 /// BEGIN bool QCircuit::validate_conditionals() const
-TEST(qpp_QCircuit_validate_condtional, AllTests) {}
+TEST(qpp_QCircuit_validate_conditionals, AllTests) {}
 
 /// BEGIN bool QCircuit::was_measured_d(idx i) const
 TEST(qpp_QCircuit_was_measured_d, AllTests) {}